feat(svm): reorganize remaining svm tools
