+++ scripts/svm/proposeRebalanceToHubPool.ts
@@ -30,14 +30,13 @@ import { getNodeUrl } from "@uma/common";
 import { BigNumber, ethers } from "ethers";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
+import { getSolanaChainId, SOLANA_USDC_DEVNET, SOLANA_USDC_MAINNET } from "../../src/svm";
 import { BondToken__factory, HubPool__factory } from "../../typechain";
 import { CHAIN_IDs } from "../../utils/constants";
-import { SOLANA_USDC_DEVNET, SOLANA_USDC_MAINNET } from "./utils/constants";
 import {
   constructEmptyPoolRebalanceTree,
   constructSimpleRebalanceTreeToHubPool,
   formatUsdc,
-  getSolanaChainId,
   requireEnv,
 } from "./utils/helpers";
