+++ scripts/svm/proposeRebalanceToSpokePool.ts
@@ -7,15 +7,16 @@
 // - NODE_URL_11155111: Ethereum RPC URL for Sepolia (ignored if TESTNET=false).
 
 // eslint-disable-next-line camelcase
-import { CHAIN_IDs, TOKEN_SYMBOLS_MAP } from "../../utils/constants";
+import { getNodeUrl } from "@uma/common";
+import { BigNumber, ethers } from "ethers";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
-import { ethers, BigNumber } from "ethers";
-import { getNodeUrl } from "@uma/common";
+import { CHAIN_IDs, TOKEN_SYMBOLS_MAP } from "../../utils/constants";
 // eslint-disable-next-line camelcase
+import { getSolanaChainId } from "../../src/svm";
 import { BondToken__factory, HubPool__factory } from "../../typechain";
+import { requireEnv } from "./utils/helpers";
 import { constructSimpleRebalanceTree } from "./utils/poolRebalanceTree";
-import { getSolanaChainId, requireEnv } from "./utils/helpers";
 
 // Set up Ethereum provider.
 const nodeURL = process.env.TESTNET === "true" ? getNodeUrl("sepolia", true) : getNodeUrl("mainnet", true);
